+++ src/Numeric/Units/Dimensional.hs
@@ -191,7 +191,7 @@ module Numeric.Units.Dimensional
     exp, log, logBase, sin, cos, tan, asin, acos, atan, sinh, cosh, tanh, asinh, acosh, atanh, atan2,
     -- ** Operations on Collections
     -- $collections
-    (*~~), (/~~), sum, mean, dimensionlessLength, nFromTo,
+    (*~~), (/~~), sum, mean, product, dimensionlessLength, nFromTo,
     -- * Dimension Synonyms
     -- $dimension-synonyms
     DOne, DLength, DMass, DTime, DElectricCurrent, DThermodynamicTemperature, DAmountOfSubstance, DLuminousIntensity,
@@ -233,7 +233,7 @@ import Data.Maybe
 import Data.Ratio
 import Numeric.Units.Dimensional.Dimensions
 import Numeric.Units.Dimensional.Internal
-import Numeric.Units.Dimensional.UnitNames hiding ((*), (/), (^), weaken, strengthen)
+import Numeric.Units.Dimensional.UnitNames hiding ((*), (/), (^), weaken, strengthen, product)
 import qualified Numeric.Units.Dimensional.UnitNames.Internal as Name
 import Numeric.Units.Dimensional.Variants hiding (type (*), type (/))
 import qualified Numeric.Units.Dimensional.Variants as V
@@ -547,7 +547,7 @@ xs /~~ u = fmap (/~ u) xs
 
 infixl 7  *~~, /~~
 
--- | The sum of all elements in a list.
+-- | The sum of all elements in a foldable structure.
 --
 -- >>> sum ([] :: [Mass Double])
 -- 0.0 kg
@@ -557,7 +557,17 @@ infixl 7  *~~, /~~
 sum :: (Num a, Foldable f) => f (Quantity d a) -> Quantity d a
 sum = foldr (+) _0
 
--- | The arithmetic mean of all elements in a list.
+-- | The product of all elements in a foldable structure.
+--
+-- >>> product ([] :: [Dimensionless Double])
+-- 1.0
+--
+-- >>> product [pi, _4, 0.36 *~ one]
+-- 4.523893421169302
+product :: (Num a, Foldable f) => f (Dimensionless a) -> Dimensionless a
+product = foldr (*) _1
+
+-- | The arithmetic mean of all elements in a foldable structure.
 --
 -- >>> mean [pi, _7]
 -- 5.070796326794897
@@ -743,8 +753,11 @@ we provide a means for converting from type-level dimensions to term-level dimen
 -- Supplying negative defining quantities is allowed and handled gracefully, but is discouraged
 -- on the grounds that it may be unexpected by other readers.
 mkUnitR :: Floating a => UnitName m -> ExactPi -> Unit m1 d a -> Unit m d a
-mkUnitR n s' (Unit _ s x) | isExactZero s = error "Supplying zero as a conversion factor is not valid."
-                          | otherwise     = Unit n (s' Prelude.* s) (approximateValue s' Prelude.* x)
+mkUnitR n s (Unit _ e _) | isExactZero s = error "Supplying zero as a conversion factor is not valid."
+                         | otherwise     = Unit n e' x'
+  where
+    e' = s Prelude.* e
+    x' = approximateValue e'
 
 -- | Forms a new atomic 'Unit' by specifying its 'UnitName' and its definition as a multiple of another 'Unit'.
 --
@@ -753,11 +766,12 @@ mkUnitR n s' (Unit _ s x) | isExactZero s = error "Supplying zero as a conversio
 --
 -- For more information see 'mkUnitR'.
 mkUnitQ :: Fractional a => UnitName m -> Rational -> Unit m1 d a -> Unit m d a
-mkUnitQ n s' (Unit _ s _) | s' == 0                       = error "Supplying zero as a conversion factor is not valid."
-                          | Just q <- toExactRational s'' = Unit n s'' (fromRational q)
-                          | otherwise                     = error "The resulting conversion factor is not an exact rational."
+mkUnitQ n s (Unit _ e x) | s == 0    = error "Supplying zero as a conversion factor is not valid."
+                         | Just x'' <- toExactRational e' = Unit n e' (fromRational x'')
+                         | otherwise = Unit n e' x'
   where
-    s'' = fromRational s' Prelude.* s
+    e' = fromRational s Prelude.* e
+    x' = fromRational s Prelude.* x
 
 -- | Forms a new atomic 'Unit' by specifying its 'UnitName' and its definition as a multiple of another 'Unit'.
 --
@@ -766,8 +780,9 @@ mkUnitQ n s' (Unit _ s _) | s' == 0                       = error "Supplying zer
 --
 -- For more information see 'mkUnitR'.
 mkUnitZ :: Num a => UnitName m -> Integer -> Unit m1 d a -> Unit m d a
-mkUnitZ n s' (Unit _ s _) | s' == 0                      = error "Supplying zero as a conversion factor is not valid."
-                          | Just z <- toExactInteger s'' = Unit n s'' (fromInteger z)
-                          | otherwise                    = error "The resulting conversion factor is not an exact integer."
+mkUnitZ n s (Unit _ e x) | s == 0    = error "Supplying zero as a conversion factor is not valid."
+                         | Just x'' <- toExactInteger e' = Unit n e' (fromInteger x'')
+                         | otherwise = Unit n e' x'
   where
-    s'' = fromInteger s' Prelude.* s
+    e' = fromInteger s Prelude.* e
+    x' = fromInteger s Prelude.* x